Remove plain-text credential store
librespot stores a more secure token that is valid for a while, rely on this
instead.

On the flip side this requires users to re-enter their login data when their
token has expired.

If the token validity is too short we will have to come up with another
approach, e.g. OS keyrings.

fixes #447