fix: panic on `SIGHUP` by closed `/dev/tty`
This fixes a panic caused by the `SIGHUP` signal being sent to the
process after its controlling terminal closes. Handling `SIGHUP` isn't
necessary as we're not interested in whether the controlling terminal
closed, only in the fact that we need to terminate. `SIGTERM` will
always be sent as well so just react to that.