chore: keep track of used toolchain version
Keeping the used toolchain version in version control makes sure that
when building older versions, the same toolchain is used every time. For
example, when `ncspot 0.12.0` has always been compiled with Rust version
`1.67.0` but the currently installed version is `1.73.0`, this file
would cause Rustup to download the older version again.