chore: remove unused `rustfmt.toml` file
This configuration file is empty and therefore doesn't do anything. It
makes sense to remove it to keep the files and directories in the root
of the project as minimal as possible. This makes it easier for new
contributors to get a quick overview of the project. The more files
there are in the root of a project, the harder it gets to tell what's
important.