fix: prevent crash when song has no artist
When loading this playlist
(https://open.spotify.com/playlist/2onE4ObADgp1NuBzciYF0Q), ncspot would
crash. The exact cause is not clear, but it seems to be caused by the
songs not being available and therefore not having a valid artist.
Therefore ensure we have a default value for artists in case one isn't
available.